+++ include/CGAL/Arr_point_location/Arr_lm_random_generator.h
@@ -12,8 +12,8 @@
 // This file is provided AS IS with NO WARRANTY OF ANY KIND, INCLUDING THE
 // WARRANTY OF DESIGN, MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE.
 //
-// $URL: svn+ssh://scm.gforge.inria.fr/svn/cgal/branches/releases/CGAL-4.1-branch/Arrangement_on_surface_2/include/CGAL/Arr_point_location/Arr_lm_random_generator.h $
-// $Id: Arr_lm_random_generator.h 70071 2012-06-26 12:52:02Z efif $
+// $URL$
+// $Id$
 // 
 // Author(s)     : Idit Haran   <haranidi@post.tau.ac.il>
 //                 Ron Wein     <wein@post.tau.ac.il>
@@ -129,7 +129,7 @@ protected:
     // Create N random landmarks. If N was not given to the constructor,
     // set it to be the number of vertices in the arrangement.
     if (num_landmarks == 0)
-      num_landmarks = arr->number_of_vertices();
+      num_landmarks = static_cast<unsigned int>(arr->number_of_vertices());
 
     CGAL::Random random;
     for (unsigned int i = 0; i < num_landmarks; ++i) {
